feat: update Reports API to improve data handling and caching, refact…

+++ ReportsAPI/Services/IDashBoardStatsService.cs
@@ -3,5 +3,5 @@ using ReportsAPI.DTOs.response;
 public interface IDashBoardStatsService
 {
     public Task<DashBoardStatsResponse> GetDashBoardStats();
-    public Task<AssessmentResponse> GetAssessmentStats(int id);
+    public Task<AssessmentResponse> GetAssessmentStats(long id);
 }
